backend/requirements.txt
fastapi[all]
sqlmodel
bcrypt
python-jose[cryptography]
psycopg2-binary
asyncpg
uvicorn
alembic
openai>=1.0.0
pydantic>=2.0.0
hypothesis
pytest
requests
python-dotenv
